move scattering-related methods to RadiosityFast.py
