+++ sparrowpy/classes/RadiosityFast.py
@@ -5,7 +5,12 @@ from sparrowpy.form_factor import form_factor
 from sparrowpy.radiosity_fast import (
     source_energy, receiver_energy, geometry)
 from sparrowpy.radiosity_fast import energy_exchange_order as ee_order
-
+try:
+    import numba
+    prange = numba.prange
+except ImportError:
+    numba = None
+    prange = range
 
 class DirectionalRadiosityFast():
     """Radiosity object for directional scattering coefficients."""
@@ -407,6 +412,124 @@ class DirectionalRadiosityFast():
             assert (self._frequencies == frequencies).all(), \
                 "Frequencies do not match"
 
+def get_scattering_data_receiver_index(
+        pos_i:np.ndarray, pos_j:np.ndarray,
+        receivers:np.ndarray, wall_id_i:np.ndarray,
+        ):
+    """Get scattering data depending on previous, current and next position.
+
+    Parameters
+    ----------
+    pos_i : np.ndarray
+        current position of shape (3)
+    pos_j : np.ndarray
+        next position of shape (3)
+    receivers : np.ndarray
+        receiver directions of all walls of shape (n_walls, n_receivers, 3)
+    wall_id_i : np.ndarray
+        current wall id to get write directional data
+
+    Returns
+    -------
+    scattering_factor: float
+        scattering factor from directivity
+
+    """
+    n_patches = pos_i.shape[0] if pos_i.ndim > 1 else 1
+    receiver_idx = np.empty((n_patches), dtype=np.int64)
+
+    for i in range(n_patches):
+        difference_receiver = pos_i[i]-pos_j
+        difference_receiver /= np.linalg.norm(
+            difference_receiver)
+        receiver_idx[i] = np.argmin(np.sum(
+            (receivers[wall_id_i[i], :]-difference_receiver)**2, axis=-1),
+            axis=-1)
+
+
+    return receiver_idx
+
+
+def get_scattering_data(
+        pos_h:np.ndarray, pos_i:np.ndarray, pos_j:np.ndarray,
+        sources:np.ndarray, receivers:np.ndarray, wall_id_i:np.ndarray,
+        scattering:np.ndarray, scattering_index:np.ndarray):
+    """Get scattering data depending on previous, current and next position.
+
+    Parameters
+    ----------
+    pos_h : np.ndarray
+        previous position of shape (3)
+    pos_i : np.ndarray
+        current position of shape (3)
+    pos_j : np.ndarray
+        next position of shape (3)
+    sources : np.ndarray
+        source directions of all walls of shape (n_walls, n_sources, 3)
+    receivers : np.ndarray
+        receiver directions of all walls of shape (n_walls, n_receivers, 3)
+    wall_id_i : np.ndarray
+        current wall id to get write directional data
+    scattering : np.ndarray
+        scattering data of shape (n_scattering, n_sources, n_receivers, n_bins)
+    scattering_index : np.ndarray
+        index of the scattering data of shape (n_walls)
+
+    Returns
+    -------
+    scattering_factor: float
+        scattering factor from directivity
+
+    """
+    difference_source = pos_h-pos_i
+    difference_receiver = pos_i-pos_j
+
+    difference_source /= np.linalg.norm(difference_source)
+    difference_receiver /= np.linalg.norm(difference_receiver)
+    source_idx = np.argmin(np.sum(
+        (sources[wall_id_i, :, :]-difference_source)**2, axis=-1))
+    receiver_idx = np.argmin(np.sum(
+        (receivers[wall_id_i, :]-difference_receiver)**2, axis=-1))
+    return scattering[scattering_index[wall_id_i],
+        source_idx, receiver_idx, :]
+
+
+def get_scattering_data_source(
+        pos_h:np.ndarray, pos_i:np.ndarray,
+        sources:np.ndarray, wall_id_i:np.ndarray,
+        scattering:np.ndarray, scattering_index:np.ndarray):
+    """Get scattering data depending on previous, current position.
+
+    Parameters
+    ----------
+    pos_h : np.ndarray
+        previous position of shape (3)
+    pos_i : np.ndarray
+        current position of shape (3)
+    sources : np.ndarray
+        source directions of all walls of shape (n_walls, n_sources, 3)
+    wall_id_i : np.ndarray
+        current wall id to get write directional data
+    scattering : np.ndarray
+        scattering data of shape (n_scattering, n_sources, n_receivers, n_bins)
+    scattering_index : np.ndarray
+        index of the scattering data of shape (n_walls)
+
+    Returns
+    -------
+    scattering_factor: float
+        scattering factor from directivity
+
+    """
+    difference_source = pos_h-pos_i
+    difference_source /= np.linalg.norm(difference_source)
+    source_idx = np.argmin(np.sum(
+        (sources[wall_id_i, :, :]-difference_source)**2, axis=-1))
+    return scattering[scattering_index[wall_id_i], source_idx]
+
+
+
+
     @property
     def n_bins(self):
         """Return the number of frequency bins."""
@@ -505,3 +628,11 @@ def _rotate_coords_to_normal(
     sources_cp.rotate('xyz', euler)
     sources_cp.radius = 1
     return sources_cp, receivers_cp
+
+
+if numba is not None:
+    get_scattering_data_receiver_index = numba.njit()(
+        get_scattering_data_receiver_index)
+    get_scattering_data = numba.njit()(get_scattering_data)
+    get_scattering_data_source = numba.njit()(get_scattering_data_source)
+
